fix(rp2040): clear PC LSB in exceptionEntry()
According to the datasheet, the actual value pushed to the stack is `ReturnAddress(ExceptionType);` (and not simply PC):

```
  MemA[frameptr+0x18,4] = ReturnAddress(ExceptionType);
```

In most cases, this equals to the address of the next instruction (the
value of PC). But there's also a comment that says:

// NOTE: ReturnAddress() is always halfword aligned, meaning bit<0> is always zero

Why is this important?

The SVC (Supervisor Call) instruction accepts a 8-bit immediate operand.

This operand may be interpreted in software, as outlined in the datasheet:

```
imm32 = ZeroExtend(imm8, 32);
// imm32 is for assembly/disassembly. SVC handlers in some
// systems interpret imm8 in software, for example to determine the required service.
```

In case of Mbed OS, the code for `SVC_Handler` indeed reads the value of the `imm8` operand.

It does so by looking at the PC value that was saved to stack in `exceptionEntry()`, subtracting 2, and then loading a byte from the resulting memory address. Here's an excerpt from the code that loads the operand of SVC into r1:

```
<SVC_Handler+6>      mrs     r0, PSP
<SVC_Handler+10>     ldr     r1, [r0, #24]
<SVC_Handler+12>     subs    r1, #2
<SVC_Handler+14>     ldrb    r1, [r1, #0]
<SVC_Handler+16>     cmp     r1, #0
```

Thus, if stack contains the value of PC with the LSB set, the code will read the operand value from the wrong address. Specifically, it'll always read 223, which is the high byte of the SVC instruction.

Clearing the LSB of PC before pushing it into the stack in `exceptionEntry()` solves this issue.
It also seems to fix the Mbed OS startup crash we encountered during the last live stream.