fix(usb): add read delay, improved CDC latency
New CDC behavior: when the code reads data from CDC and the Tx FIFO is empty, we just return an empty buffer instead of waiting until there's some data to send.

Otherwise, this causes output latency in CircuitPython. Also, tested it against the hardware and it seem to be sending empty buffers when there's nothing to send over CDC, matching the new behavior.