fix(rp2040): WFI doesn't wake up when interrupts disabled
according to the ARM ARM, we should wake even if PM=1, as long as the exception has higher priority than the currently executing handler:

> If PRIMASK.PM is set to 1, an asynchronous exception that has a higher group priority than any
active exception results in a WFI instruction exit. If the group priority of the exception is less than or
equal to the execution group priority, the exception is ignored.

see adafruit/circuitpython#5331