feat: simulate a given UF2 firmware using `npm start` (#143)
The existing implementtaion of `start` does not accept
a path on the command line while it has the feature of
relaying UART0 to the console.

Add the command line parameter to specify an arbitrary image,
also accept UF2 file, besides hex, preserve backward compat.
As a result, images that send output to UART0 can be benefit
from that.

Co-authored-by: kromych <kromych@users.noreply.github.com>